Moved document window stuff into DocumentWindowController. Added dynamic row height based on contents of bubble's body property. Added tableview reload on resize so that row heights will match the height of the wrapped text.
